+++ packages/toprf-secure-backup/src/interfaces.ts
@@ -128,6 +128,7 @@ export type CreateLocalKeyResult = {
   seed: Uint8Array;
   authKeyPair: KeyPair;
   encKey: Uint8Array;
+  pwEncKey: Uint8Array;
 };
 
 /**
@@ -187,13 +188,19 @@ export type CreateEncryptionKeyParams = {
  * authKeyPair - The authentication key pair which is used to authenticate the user.
  *
  * encKey - The encryption key which is used to encrypt the secret data.
+ *
+ * pwEncKey - The password encryption key which is used to encrypt the password.
  */
 export type CreateEncryptionKeyResult = {
   authKeyPair: KeyPair;
   encKey: Uint8Array;
+  pwEncKey: Uint8Array;
 };
 
-export type BaseAddSecretDataItemParams<SecretDataType> = {
+export type BaseAddSecretDataItemParams<
+  SecretDataType,
+  EncKeyType = Uint8Array,
+> = {
   /**
    * The secret data to be stored.
    */
@@ -202,7 +209,7 @@ export type BaseAddSecretDataItemParams<SecretDataType> = {
   /**
    * The encryption key to be used to encrypt the secret data.
    */
-  encKey: Uint8Array;
+  encKey: EncKeyType;
 
   /**
    * The authentication key to be used to provide valid signature for storing the secret data.
@@ -220,7 +227,8 @@ export type BaseAddSecretDataItemParams<SecretDataType> = {
 export type AddSecretDataItemParams = BaseAddSecretDataItemParams<Uint8Array>;
 
 export type BatchAddSecretDataItemParams = BaseAddSecretDataItemParams<
-  Uint8Array[]
+  Uint8Array[],
+  Uint8Array | Uint8Array[]
 >;
 
 /**
@@ -253,6 +261,7 @@ export type RecoverEncryptionKeyParams = {
 export type RecoverEncryptionKeyResult = {
   authKeyPair: KeyPair;
   encKey: Uint8Array;
+  pwEncKey: Uint8Array;
   keyShareIndex: number;
   rateLimitResetResult: Promise<void>;
 };
@@ -270,8 +279,6 @@ export type RecoverEncryptionKeyResult = {
  *
  * oldAuthKeyPair - The old authentication key pair of the user.
  *
- * oldPassword - The old password of the user.
- *
  * newPassword - The new password of the user.
  *
  * newKeyShareIndex - The key share index to be used for the new key.
@@ -283,8 +290,8 @@ export type ChangeEncryptionKeyParams = {
   authConnectionId: string;
   userId: string;
   oldEncKey: Uint8Array;
+  oldPwEncKey: Uint8Array;
   oldAuthKeyPair: KeyPair;
-  oldPassword: string;
   newPassword: string;
   newKeyShareIndex: number;
   groupedAuthConnectionId?: string;
@@ -297,6 +304,7 @@ export type ChangeEncryptionKeyParams = {
 export type ChangeEncryptionKeyResult = {
   authKeyPair: KeyPair;
   encKey: Uint8Array;
+  pwEncKey: Uint8Array;
 };
 
 /**
@@ -345,15 +353,15 @@ export type FetchAuthPubKeyResult = {
   authPubKey: SEC1EncodedPublicKey;
 };
 
-export type RecoverPasswordParams = {
-  targetPwPubKey: SEC1EncodedPublicKey;
-  curEncKey: Uint8Array;
+export type RecoverPwEncKeyParams = {
+  targetAuthPubKey: SEC1EncodedPublicKey;
+  curPwEncKey: Uint8Array;
   curAuthKeyPair: KeyPair;
   maxPwChainLength?: number;
 };
 
-export type RecoverPasswordResult = {
-  password: string;
+export type RecoverPwEncKeyResult = {
+  pwEncKey: Uint8Array;
 };
 
 export type IToprfSecureBackup = {
@@ -466,18 +474,18 @@ export type IToprfSecureBackup = {
   ) => Promise<FetchAuthPubKeyResult>;
 
   /**
-   * This function recovers the password of the user.
+   * This function recovers the password encryption key of the user.
    *
-   * @param params - The parameters for recovering the password.
-   * @param params.targetPwPubKey - The public key of the target password.
-   * @param params.curEncKey - The current encryption key of the user.
+   * @param params - The parameters for recovering the password encryption key.
+   * @param params.targetAuthPubKey - The public key of the target encryption key.
+   * @param params.curPwEncKey - The current password encryption key of the user.
    * @param params.curAuthKeyPair - The current authentication key pair of the user.
    *
-   * @returns A promise that resolves with the password of the user.
+   * @returns A promise that resolves to the password encryption key of the user.
    */
-  recoverPassword: (
-    params: RecoverPasswordParams,
-  ) => Promise<RecoverPasswordResult>;
+  recoverPwEncKey: (
+    params: RecoverPwEncKeyParams,
+  ) => Promise<RecoverPwEncKeyResult>;
 };
 
 /**